Enable out-of-source-tree builds
One can now do:

    mkdir build-tree && cd build-tree
    ../user-manual/configure --with-tor-browser-bundle=../tor-browser-bundle
    make html

Easy part was sprinkling Makefile.am with appropriate $(srcdir). Symlinks
to the media needs to be done to localized version or to C version as a
fallback.

Less funny was how to deal with <xi:include /> for the version number.
We use the `-p` option of yelp-build to specify a search path. Works great.
But there's no such option to `itstool` and it errors out if it's unable
to find `type="text"` includes (to get include them in translation strings, I
presume). So we now include the Tor Browser version in an XML file instead.

Gotcha: yelp rules uses the presence of the C directory to detect if it's
an out-of-tree build. So we must not create a C directory when building.

Let's also re-add a missing variable substitution in configure.ac.